+++ my-app/.env
@@ -4,7 +4,7 @@
 # Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
 # See the documentation for all the connection string options: https://pris.ly/d/connection-strings
 
-DATABASE_URL="mongodb+srv://visarnlerdsiri:59UqUMhwg5ZaujG5@project.et14qwc.mongodb.net/project?retryWrites=true&w=majority&appName=Project"
+DATABASE_URL="mongodb+srv://admin:CVS23snwBoeORjGQ@cluster0.pnmzxd6.mongodb.net/myDatabase"
 NEXTAUTH_URL="http://localhost:3000"
 NEXTAUTH_SECRET="dLmzvISXAHi47ztdEjYfSKfMQ2+SeK24vbE9fas6HVkmETJ/DnR79hEr6mbY3id1M+PVuk0W8LUFls8c+FsIMcKtkItCpnA2ba190DT15nJlZtlJr/uOjjmTc/4UiQBzae5JX4uHpw7stPdwaM0gJmlxDjgJ/gP/UoRYILA8zLTxdUee0XphGAlzfUy1IY2r"
 NEXT_MAPS_API_KEY="AIzaSyC6LYgvdJqZ0QLoJXA7XKLHuaqPPzLY1Ac"
