+++ my-app/app/(user)/map/page.tsx
@@ -1,4 +1,3 @@
-
 'use client';
 
 import React from 'react'
@@ -148,6 +147,5 @@ export default function page() {
 
   )
 }
-
