+++ my-app/app/(user)/profile/myplan/page.tsx
@@ -1,4 +1,3 @@
-
 import Link from "next/link";
 import { FaUserCircle } from "react-icons/fa"
 import { IoSettingsOutline } from "react-icons/io5"
@@ -49,4 +48,3 @@ export default function Home() {
     );
 
 }
-
