+++ my-app/app/components/Maps.tsx
@@ -14,8 +14,7 @@ export default function Maps() {
     useEffect(() =>{
         const initMap = async () => {
             const loader = new Loader({
-                apiKey: process.env.NEXT_PUBLIC_MAPS_API_KEY!,
-
+                apiKey: "AIzaSyC6LYgvdJqZ0QLoJXA7XKLHuaqPPzLY1Ac",
                 version:'weekly',
                 libraries:['places']
             }) 
@@ -65,6 +64,8 @@ export default function Maps() {
                 }
               }
             }
+
+            
     }
         initMap();
     }, [])
