--- my-app/app/components/profile/accountset.tsx
@@ -1,7 +1,0 @@
-export default function Accountset() {
-    return(
-        <div className="card max-w-full bg-[#D3BD9A] rounded-md px-10 md:px-20 py-20">
-            account set
-        </div>
-    )
-}
