+++ my-app/app/types/next-auth.d.ts
@@ -8,7 +8,6 @@ declare module "next-auth" {
     interface Session {
         user: {
         /** The user's postal address. */
-            id: string,
             username: string,
             email: string,
             role: roles
