+++ my-app/package.json
@@ -16,17 +16,14 @@
   "dependencies": {
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
-    "@googlemaps/js-api-loader": "^1.16.6",
     "@headlessui/react": "^1.7.18",
     "@heroicons/react": "^2.1.1",
     "@mantine/core": "^7.6.1",
     "@mantine/dates": "^7.6.1",
     "@mantine/hooks": "^7.6.1",
     "@mui/material": "^5.15.11",
-    "@prisma/client": "^5.11.0",
-    "@react-google-maps/api": "^2.19.3",
+    "@prisma/client": "^5.9.0",
     "@types/bcrypt": "^5.0.2",
-    "@types/google.maps": "^3.55.5",
     "bcrypt": "^5.1.1",
     "date-fns": "^3.3.1",
     "dayjs": "^1.11.10",
@@ -55,7 +52,7 @@
     "eslint": "^8",
     "eslint-config-next": "14.0.4",
     "postcss": "^8",
-    "prisma": "^5.11.0",
+    "prisma": "^5.10.2",
     "tailwindcss": "^3.3.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.3"
