+++ my-app/prisma/schema.prisma
@@ -13,11 +13,10 @@ datasource db {
 model user {
   id       String    @id @default(auto()) @map("_id") @db.ObjectId
   username String    @unique
-  displayName String? 
   email    String    @unique
   password String
-  role     roles  @default(user)
-  plan     plan[]
+  role     roles     @default(user)
+  plan     planAll[]
 }
 
 enum roles {
@@ -26,30 +25,27 @@ enum roles {
   superadmin
 }
 
-model plan {
-  id        String     @id @default(auto()) @map("_id") @db.ObjectId
-  name      String?    @default("My plan")
-  budget    Int?       @default(0)
-  images    String?
-  detail    String?
-  startDate DateTime
-  endDate   DateTime
-  owner     user       @relation(fields: [ownerId], references: [id])
-  ownerId   String     @db.ObjectId
-  status    planStatus @default(draft)
+model planning {
+  id        String   @id @default(auto()) @map("_id") @db.ObjectId
+  name      String
+  budget    Int
+  images    String
+  date      DateTime
+  plan      planAll  @relation(fields: [planAllId], references: [id])
+  planAllId String   @db.ObjectId
 }
 
-enum planStatus {
-  public
-  draft
-}
+model planAll {
+  id         String   @id @default(auto()) @map("_id") @db.ObjectId
+  name       String
+  tripBudget Int
+  images     String
+  startDate  DateTime
+  endDate    DateTime
+  owner      user     @relation(fields: [ownerId], references: [id])
+  ownerId    String   @db.ObjectId
 
-model touristAttraction {
-  id               String  @id @default(auto()) @map("_id") @db.ObjectId
-  name             String?
-  images           String?
-  googleCoodinates Json
-  googlePlaceId    String?
+  planByDay planning[]
 }
 
 model location {
